Refactored the code and removed debug statements
Refactored the code to make it easier to understand.
Split the declaration and Implementation of the disk class into header and source files to ease maintenance.
Still generating seg faults when evolving particles.
