+++ freeze_thaw.cpp
@@ -1,12 +1,10 @@
 // freeze_thaw.cpp : Defines the entry point for the console application.
-//
-
+//************************************************************************//
+//*******************************Inclusions*******************************//
+//************************************************************************//
 #include "stdafx.h"
 #include "disk.h"
 #include "Vector.h"
-//#include "init_hopper.h"
-//#include "init_silo.h"
-//#include "init_window_box.h"
 #include<iostream>
 #include<fstream>
 #include<cstdlib>
@@ -21,16 +19,24 @@
 #include<stdio.h>
 #include<limits>
 
+//************************************************************************//
+//*********************************Defines********************************//
+//************************************************************************//
 #define length(a) ( sizeof ( a ) / sizeof ( *a ) )//Length of an array
 //#define DEBUG
 using namespace std;
 
-//Function Headers
+//************************************************************************//
+//****************************Function Headers****************************//
+//************************************************************************//
+//Assign the system properties from a file
+void assign_system_properties(istream & in);
+//Assign the system properties from a file
+void assign_particle_properties(istream & in);
 //Dump particle properties to the console
 void dump_grain_to_console(int i, disk p);
 //Dump particle properties to file
 void dump_grain_to_file(ostream & os, int i, disk p);
-
 //Calculate the packing density of the system
 double density();
 //Calculate the Total Kinetic Energy of the System
@@ -41,16 +47,24 @@ double total_linear_momentum();
 double total_angular_momentum();
 //Calculate the Average Coordination Number of the System
 double avg_coordination_number();
+//Calculate the time taken to run the simulation
+void calc_execution_time(time_t startTime, time_t endTime);
 //The Verlet List functions
 bool make_verlet();
 bool do_touch(int i, int k);
 bool verlet_needs_update();
-void predict(double dt);
-void correct(double dt);
+void resize_cells(vector<disk> &particle, vector<vector<vector<int> > > &celllist);
 
+//************************************************************************//
+//****************************System Variables****************************//
+//************************************************************************//
 int particles;						//The number of particles in the system
 int tag;							//Used to select the force law
 int output;							//Used to dump the phase properties
+double dt,tmax, totalSteps;			//The timestep,max time and total steps in the simulation
+time_t start, end;					//Timer variables
+double diff;						//The difference in time
+int counter = 0;					//Print out counter
 
 Vector G;							//The Gravity vector
 vector<disk> particle;				//The vector that holds all particle objects
@@ -59,7 +73,10 @@ double potential_sum = 0.0;			//The sum of the system potential energy
 double kinetic_sum = 0.0;			//The sum of the system kinetic energy
 double linear_momentum_sum = 0.0;	//The sum of the system linear momentum
 double angular_momentum_sum = 0.0;	//The sum of the system angular momentum
-//Verlet variables
+
+//************************************************************************//
+//****************************Verlet Variables****************************//
+//************************************************************************//
 vector<set<int> > verlet;			//The Verlet List
 vector<vector<vector<int> > > celllist;//The Grid List
 vector<disk> safe;					//Backup of the vector that holds all particle objects
@@ -76,55 +93,146 @@ double x_0 = 0.0;
 double y_0 = 0.0;
 double Timesafe;
 
+//************************************************************************//
+//**********************************Main**********************************//
+//************************************************************************//
+
 int main(int argc, char* argv[])//Main program
 {	
+	//Open files for input/output
+	ofstream out1;					//File writer object, writes phase information at predetermined timesteps
+	out1.open("Output/run1.bin",ios::out | ios::binary);
 
-	//Init variables
-	double dt,tmax, totalSteps;		//The timestep and the max time
-	time_t start, end;	//Timer variables
-	double diff;		//The difference in time
-	int j;			//Counters for loops
-	int counter = 0;	//Print out counter
-
-    //Open files for input/output
-	ofstream out1 ("Output/run1.out");	//File writer object, writes phase information at predetermined timesteps
-	ofstream out2 ("Input/run2.in");	//File writer object, writes the final phase information at the end
-	//ifstream in ("input.dat");//File reader object
-	ifstream in ("Input/hopper.in");//File reader object, change this to run any tests
-    #ifdef DEBUG
-	    //Testing output file, this is for any information
-	    ofstream vel ("Output/vel2.out");
-
-	    //Output file for the energy of the system
-	    ofstream eng ("Output/eng2.out");
-
-	    //Output file for the Linear Momentum of the system
-	    ofstream lmn ("Output/lmn2.out");
-
-	    //Output file for the Angular Momentum of the system
-	    ofstream amn ("Output/amn2.out");
-
-	    //Output file for the Coordination Numberof the system
-	    ofstream cdn ("Output/cdn.out");
-    #endif
+	ifstream in;//File reader object, change this to run any tests
+	in.open("Input/grv_test.in",ios::in);
 
 	if(!out1){//If the file is not accessible
-		cout << "Cannot open the first file to be written to!\n";
-		return 1;
-	}
-
-	if(!out2){//If the file is not accessible
-		cout << "Cannot open the seccond file to be written to!\n";
+		cerr << "Cannot open the first file to be written to!\n";
 		return 1;
 	}
 
 	if(!in){//If the file is not present
-		cout << "Cannot access the input file!\n";
+		cerr << "Cannot access the input file!\n";
 		return 1;
 	}
+	int j;								//Counters for loops
+
+	assign_system_properties(in);
+
+	//Print Pre-processing details to the screen
+    totalSteps = tmax/dt;
+	output = int(totalSteps/100);
+	cout << "The Simulation is of " << particles << " particles" << ".\n" 
+		<< "It simulates "<< tmax << " second(s) at a time step of " << dt << ".\n"
+		<< "Using force law " << tag << ".\n"
+	    << "This equates to " << totalSteps << " steps. \n"
+	    << "Every " << output << " steps there will be a file write operation.\n";
 	
-	cout << "The input file is being read:\n";
-	//Read data from file, Gravity
+	resize_cells(particle,celllist);
+
+	assign_particle_properties(in);
+
+	//Close the input files
+	in.close();
+	make_verlet();
+
+	kinetic_sum=total_kinetic_energy();
+	linear_momentum_sum = total_linear_momentum();
+	angular_momentum_sum = total_angular_momentum();
+
+	
+	//Particle Properties, goes to run1.out
+	for(unsigned int j = 0; j < particles; j++){
+		dump_grain_to_file(out1,j,particle[j]);
+	}
+	
+	//Start time
+	time(&start);
+	//Print it to screen
+	cout << "The Main loop has started:\n";
+	for(double t = 0; t <= tmax; t+=dt){
+		counter++;
+		bool ok = true;
+		bool newverlet = false;
+		//Check if the verlet list must be updated
+		if(verlet_needs_update()){
+			ok = make_verlet();
+			newverlet = true;
+		}
+		//If error detected, roll back to the last known correct positions and remake the verlet list.
+		if(!ok){
+			particle = safe;
+			t = Timesafe;
+			verlet_distance *= verlet_increase;
+			make_verlet();
+		}
+		//Roll the simulation back to the last correct state if needed
+		if(newverlet && ok){
+			safe = particle;
+			Timesafe = t;
+		}
+		//Predict the new properties of the free particles, or apply boundary conditions
+		for(unsigned int i = 0; i < particle.size(); i++){
+			particle[i].reset_coordination_number();
+			if(particle[i].type()==0){
+			    particle[i].set_force_to_zero();
+			    particle[i].predict(dt);
+			}
+			else {
+					particle[i].boundary_conditions(i,dt,t);
+				 }
+		}
+		//Cycle through the particles and look for collisions
+		for(unsigned int i = 0; i < particle.size(); i++){
+			set<int>::iterator iter;
+			for(iter = verlet[i].begin(); iter != verlet[i].end(); iter++){
+				force(particle[i],particle[*iter],lx,ly,tag);
+				}
+			}
+		//Calculate the average Coordination number of the system
+		Z = avg_coordination_number();
+
+		//Correct the positions and reset the Coordination Numbers
+		for(unsigned int i = 0; i < particle.size(); i++) {
+			if(particle[i].type()==0){
+			    particle[i].correct(dt);
+			}
+			particle[i].reset_coordination_number();
+			//The system kinetic energy
+			kinetic_sum = total_kinetic_energy();
+			//The system linear momentum
+			linear_momentum_sum = total_linear_momentum();
+			//The system angular momentum
+			angular_momentum_sum = total_angular_momentum();
+		}
+
+		//Output to File, goes to run1.out
+		if(counter%output==0){
+			cout << float(counter/totalSteps)*100.0 << " percent complete \n";
+			for(j = 0; j < particles; j++){
+			    dump_grain_to_file(out1,j,particle[j]);
+		    }
+		}
+		kinetic_sum=potential_sum=0.0;
+	}
+	//End Time
+	time(&end);
+	
+	//Calculate the time to run the loop
+	calc_execution_time(start,end);
+
+	//Close the output files
+	out1.close();
+
+	return 0;
+}
+
+//************************************************************************//
+//********************************Functions*******************************//
+//************************************************************************//
+
+void assign_system_properties(istream & in){
+	cout << "Assigning the system properties...";
 	in >> particles;
 	in >> tag;
 	in >> tmax;
@@ -132,38 +240,11 @@ int main(int argc, char* argv[])//Main program
 	in >> G.x();
 	in >> G.y();
 	in >> G.phi();
+	cout << "OK\n";
+}
 
-	output = 10000;
-
-	//Print Pre-processing details to the screen
-    totalSteps = tmax/dt;
-	cout << "The Simulation is of " << particles << " particles" << ".\n" 
-		<< "It simulates "<< tmax << " second(s) at a time step of " << dt << ".\n"
-		<< "Using force law " << tag << ".\n"
-	    << "This equates to " << totalSteps << " steps. \n";
-	
-	particle.resize(particles);
-	safe=particle;
-	Timesafe=0.0;
-	vnx = int(lx/verlet_grid);
-	vny = int(ly/verlet_grid);
-	if(vnx == 0){
-		vnx = 1;
-	}
-	if(vny == 0){
-		vny = 1;
-	}
-	dx = lx/vnx;
-	dy = ly/vny;
-
-	celllist.resize(vnx);
-	for(int i = 0; i < vnx; i++){
-		celllist[i].resize(vny);
-	}
-	make_verlet();
-
-	cout << "The particles are being assigned their properties:\n";
-	//Input the particle properties from a file
+void assign_particle_properties(istream & in){
+	cout << "The particles are being assigned their properties...";
 	for(unsigned int i = 0; i < particle.size(); i++){
 		in >> particle[i].x();
 		in >> particle[i].y();
@@ -180,213 +261,9 @@ int main(int argc, char* argv[])//Main program
 		in >> particle[i].A();
 		particle[i].J() = 0.0000000001;
 	}
-
-	//Close the input files
-	in.close();
-	make_verlet();
-
-	kinetic_sum=total_kinetic_energy();
-	linear_momentum_sum = total_linear_momentum();
-	angular_momentum_sum = total_angular_momentum();
-
-	
-	//Particle Properties, goes to run1.out
-	for(unsigned int j = 0; j < particles; j++){
-		dump_grain_to_file(out1,j,particle[j]);
-	}
-    #ifdef DEBUG
-	//Coordination Number
-	    cdn << Z << "\n";
-	
-	//Testing output file
-	    vel << particle[0].vx() << "\t" << particle[0].vy() << "\t" << particle[0].omega() << "\n";
-	
-	//Print the system energy
-	    eng << kinetic_sum << "\t" << potential_sum << "\t" << kinetic_sum + potential_sum << "\n";
-
-	//Print the system linear momentum
-	    lmn << particle[0].linear_momentum() << "\t" << particle[1].linear_momentum() << "\t" << linear_momentum_sum << "\n";
-
-	//Print the system linear momentum
-	    amn << particle[0].angular_momentum() << "\t" << particle[1].angular_momentum() << "\t" << angular_momentum_sum << "\n";
-    #endif
-	//make verlet
-	//Start time
-	time(&start);
-	//Print it to screen
-	cout << "The Main loop has started:\n";
-	for(double t = 0; t <= tmax; t+=dt){
-        #ifdef DEBUG
-		    cout << "The time is: " << t << "\n";
-        #endif
-		counter++;
-        #ifdef DEBUG
-		    cout << "Counter increased to..." << counter << "\n";
-        #endif
-		bool ok = true;
-		bool newverlet = false;
-		if(verlet_needs_update()){
-            #ifdef DEBUG
-			    cout << "(1)Verlet needs update...\n";
-            #endif
-			ok = make_verlet();
-			#ifdef DEBUG
-			    cout << "(1a)...\n";
-			#endif
-			newverlet = true;
-			#ifdef DEBUG
-			    cout << "(1b)...\n";
-			#endif
-		}
-		if(!ok){
-			#ifdef DEBUG
-			    cout << "(2)Verlet fail! Reneging...\n";
-			#endif
-			particle = safe;
-			#ifdef DEBUG
-			    cout << "(2a)...\n";
-			#endif
-			t = Timesafe;
-			#ifdef DEBUG
-			    cout << "(2b)...\n";
-			#endif
-			verlet_distance *= verlet_increase;
-			#ifdef DEBUG
-			    cout << "(2c)...\n";
-			#endif
-			make_verlet();
-			#ifdef DEBUG
-			    cout << "(2d)...\n";
-			#endif
-		}
-		if(newverlet && ok){
-			#ifdef DEBUG
-			    cout << "(3) Everything is ok...\n";
-			#endif
-			safe = particle;
-			#ifdef DEBUG
-			    cout << "(3a)...\n";
-			#endif
-			Timesafe = t;
-			#ifdef DEBUG
-			    cout << "(3b)...\n";
-			#endif
-		}
-		#ifdef DEBUG
-		    cout <<"Predicting...\n";
-		#endif
-		for(unsigned int i = 0; i < particle.size(); i++){
-			particle[i].reset_coordination_number();
-			if(particle[i].type()==0){
-			    particle[i].set_force_to_zero();
-				#ifdef DEBUG
-			        cout << "Predict...\n";
-			        dump_grain_to_console(i,particle[i]);
-				#endif
-			    particle[i].predict(dt);
-				#ifdef DEBUG
-			        dump_grain_to_console(i,particle[i]);
-				#endif
-			}
-			else {
-					particle[i].boundary_conditions(i,dt,t);
-				 }
-			#ifdef DEBUG
-			    dump_grain_to_file(out1,i,particle[i]);
-			#endif
-		}
-
-		//cout << "Cycling...\n";
-		//Cycle through the particles and look for collisions
-		for(unsigned int i = 0; i < particle.size(); i++){
-			set<int>::iterator iter;
-			for(iter = verlet[i].begin(); iter != verlet[i].end(); iter++){
-				force(particle[i],particle[*iter],lx,ly,tag);
-				}
-			}
-		//Calculate the average Coordination number of the system
-		Z = avg_coordination_number();
-
-		//Correct the positions and reset the Coordination Numbers
-		//cout << "Correcting...\n";
-		for(unsigned int i = 0; i < particle.size(); i++) {
-			if(particle[i].type()==0){
-				#ifdef DEBUG
-				    cout << "Correct...\n";
-				    dump_grain_to_console(i,particle[i]);
-				#endif
-			    particle[i].correct(dt);
-				#ifdef DEBUG
-			        dump_grain_to_console(i,particle[i]);
-			        dump_grain_to_file(out1,i,particle[i]);
-				#endif
-			}
-			particle[i].reset_coordination_number();
-			//The system kinetic energy
-			kinetic_sum = total_kinetic_energy();
-			//The system linear momentum
-			linear_momentum_sum = total_linear_momentum();
-			//The system angular momentum
-			angular_momentum_sum = total_angular_momentum();
-
-		}
-
-		//Output to File, goes to run1.out
-		if(counter%output==0){
-			cout << float(counter/totalSteps)*100.0 << " percent complete \n";
-			//cout << "The time is: " << t << "\n";
-		for(j = 0; j < particles; j++){
-			dump_grain_to_file(out1,j,particle[j]);
-		}
-        #ifdef DEBUG
-			//Coordination Number
-			cdn << Z << "\n";
-
-			//Testing output file
-			vel << particle[0].vx() << "\t" << particle[0].vy() << "\t" << particle[0].omega() << "\n";
-
-			//Print the system energy
-			eng << kinetic_sum << "\t" << potential_sum << "\t" << kinetic_sum + potential_sum << "\n";
-
-			//Print the System linear momentum
-			lmn << particle[0].linear_momentum() << "\t" << particle[1].linear_momentum() << "\t" << linear_momentum_sum << "\n";
-
-			//Print the System angular momentum
-			amn << particle[0].angular_momentum() << "\t" << particle[1].angular_momentum() << "\t" << angular_momentum_sum << "\n";
-        #endif
-		}
-		kinetic_sum=potential_sum=0.0;
-	}
-	//End Time
-	time(&end);
-	
-	//Final Phase Plot, goes to run2.out
-	/*for(j = 0; j < particles; j++){
-		dump_grain_to_file(out2,j,particle[j]);
-	}*/
-	//Calculate the time to run the loop
-	diff = difftime(end,start);
-	cout << "It has taken " << diff/60 << " minutes to run the main loop.\n";
-
-	//Calculate the system density
-	//double rho = density();
-	//print it to the output file
-	//out << "The system density is: " << rho << "\n";
-
-	out1.close();
-	out2.close();
-    #ifdef DEBUG
-	    vel.close();
-	    eng.close();
-	    lmn.close();
-	    amn.close();
-	    cdn.close();
-    #endif
-
-	return 0;
+	cout<< "OK\n";
 }
 
-//--------------------------------------------Functions-----------------------------------------
 void dump_grain_to_console(int i, disk p){
 	cout << i << ", " <<
 			p.x() <<
@@ -424,8 +301,6 @@ void dump_grain_to_file(ostream & os, int i, disk p){
 double density (){
 	double sum;	//The sum of the disk areas, the system width and height
 	double result;
-	
-			
 	sum = 0.0;
 
 	//Calculate the sum of the disk areas, from r^2 * PI
@@ -437,10 +312,6 @@ double density (){
 	return result;
 }
 
-//The kinetic energy of a particle
-double disk::kinetic_energy() const{
-	return _m*(rtd1.x()*rtd1.x()/2 + rtd1.y()*rtd1.y()/2) + _J*rtd1.phi()*rtd1.phi()/2;
-}
 //The total kinetic energy of the system
 double total_kinetic_energy()
 {
@@ -453,11 +324,6 @@ double total_kinetic_energy()
   return sum;
 }
 
-//The Linear Momentum of a particle
-double disk::linear_momentum() const{
-	return _m*(rtd1.x() + rtd1.y()) + _J*rtd1.phi();
-}
-
 //The System Linear Momentum
 double total_linear_momentum(){
 	double sum= 0.0;
@@ -468,11 +334,6 @@ double total_linear_momentum(){
 		}
 	}
 	return sum;
-}
-
-//The Angular Momentum of a particle
-double disk::angular_momentum() const{
-	return _J*rtd1.phi();
 }
 
 //The System Angular Momentum
@@ -496,43 +357,15 @@ double avg_coordination_number(){
 		sum+=particle[a].z();
 	}
 
-	result = sum/particles;
+	result = double(sum/particles);
 	return result;
 }
 
-//Gears Integration Scheme - Predict
-void disk::predict(double dt){
-			double a1 = dt;
-			double a2 = a1 * dt/2;
-			double a3 = a2 * dt/3;
-			double a4 = a3 * dt/4;
-
-			rtd0 += a1*rtd1 + a2*rtd2 + a3*rtd3 + a4*rtd4;
-			rtd1 += a1*rtd2 + a2*rtd3 + a3*rtd4;
-			rtd2 += a1*rtd3 + a2*rtd4;
-			rtd3 += a1*rtd4;
-		}
-
-void disk::correct(double dt){
-			static Vector accel, corr;
-
-			double dt_recip = 1/dt;
-
-			const double coeff0 = double(19)/double(90)*(dt*dt/double(2));
-			const double coeff1 = double(3)/double(4)*(dt/double(2));
-			const double coeff3 = double(1)/double(2)*(double(3)*dt_recip);
-			const double coeff4 = double(1)/double(12)*(double(12)*(dt_recip*dt_recip));
-
-			accel=Vector((((1/_m)*_force.x())+G.x()),(((1/_m)*_force.y())+G.y()),(((1/_J)*_force.phi())+G.phi()));
-			
-			corr=accel-rtd2;
-
-			rtd0 += coeff0*corr;
-			rtd1 += coeff1*corr;
-			rtd2 = accel;
-			rtd3 += coeff3*corr;
-			rtd4 += coeff4*corr;
-		}
+void calc_execution_time(time_t startTime,time_t endTime){
+	//Calculate the time to run the loop
+	diff = difftime(endTime,startTime);
+	cout << "It has taken " << diff/60 << " minutes to run the main loop.\n";
+}
 
 void force(disk & p1, disk & p2, double lx, double ly, int tag){
 			double dx=normalise(p1.x()-p2.x(),lx);
@@ -599,107 +432,26 @@ void force(disk & p1, disk & p2, double lx, double ly, int tag){
 		}
 }
 
-void disk::boundary_conditions(int n, double dt, double t)
-{
-  switch(type()){
-  case(0): break; //Free particle
-  case(1): break; //Wall particle
-  case(2): {
-      x()=0.5-0.4*cos(10*t);
-      y()=0.1;
-      vx()=10*0.4*sin(t);
-      vy()=0;
-    } break;
-  case(3): {
-      double xx=x()-0.5;
-      double yy=y()-0.5;
-      double xp=xx*cos(dt)-yy*sin(dt);
-      double yp=xx*sin(dt)+yy*cos(dt);
-
-      x()=0.5+xp;
-      y()=0.5+yp;
-      vx()=-yp;
-      vy()= xp;
-      omega()=1;
-    } break;
-  case(4): {
-      x()=0.5+0.1*cos(t) + 0.4*cos(t+2*n*3.14159265/128);
-      y()=0.5+0.1*sin(t) + 0.4*sin(t+2*n*3.14159265/128);
-      vx()=-0.1*sin(t) - 0.4*sin(t+2*n*3.14159265/128);
-      vy()= 0.1*cos(t) - 0.4*cos(t+2*n*3.14159265/128);
-      omega()=1;
-    } break;
-  case(5): {//Vertical Vibrating Particle
-	  y()=y()+((0.02*sin(30*t))*0.1);
-      vx()=0;
-      vy()=0.02*30*cos(30*t);
-    } break;
-  case(6): {
-      int i=n/2;
-      y()=i*0.02+0.1+0.02*sin(30*t);
-      vx()=0;
-      vy()=0.02*30*cos(30*t);
-    } break;
-  case(7): {//Expander
-	  r()+=0.001;
-	}break;
-  case(8): {//Contractor
-	  r()-=0.001;
-	}break;
-  default: {
-      cout << "ptype: " << type() << " not implemented\n";
-      abort();
-    }
-  }
-}
-
 bool make_verlet(){
 	bool ok = true;
-    #ifdef DEBUG
-	    cout << "Making the new Verlet list...\n";
-    #endif
-
 	verlet.resize(particles);
-    #ifdef DEBUG
-	    cout << "Resized...\n";
-    #endif
+
 	for(int ix = 0;ix < vnx; ix++){
 		for(int iy = 0; iy < vny; iy++){
 			celllist[ix][iy].clear();
 		}
 	}
-    #ifdef DEBUG
-	    cout << "Celllist cleared...\n";
-    #endif
 	for(unsigned int i = 0; i < particle.size(); i++){
-        #ifdef DEBUG
-		    cout << i << " " << particle[i].x() << " " << particle[i].y() << " " << x_0 << " " << y_0 << " " << dx << " " << dy <<"...\n";
-        #endif
 		int ix = int((particle[i].x()-x_0)/dx);
 		int iy = int((particle[i].y()-y_0)/dy);
-        #ifdef DEBUG
-		    if (ix < 0 || iy < 0){
-			    cout << "Negative indices detected!!! " << particle[i].type() << " " << "ix: " << ix << ", iy: " << iy << "...\n";
-		    }
-		#endif
-        #ifdef DEBUG
-		    cout << ix << "..." << iy << "...\n";
-        #endif
 		celllist[ix][iy].push_back(i);
-        #ifdef DEBUG
-		    cout << "Remaking...\n";
-        #endif
 	}
-    #ifdef DEBUG
-	    cout << "Cells remade...\n";
-	#endif
 	for(unsigned int i = 0; i < particles; i++){
 		set<int> oldverlet=verlet[i];
 		verlet[i].clear();
 		int ix = int((particle[i].x()-x_0)/dx);
 		int iy = int((particle[i].y()-y_0)/dy);
 		if (ix < 0 || iy < 0){
-			cout << "Negative indices detected in make_verlet!!!" << "ix: " << ix << ", iy: " << iy;
 		}
 		for(int iix = ix - 1; iix <= ix + 1; iix++){
 			for(int iiy = iy - 1; iiy <= iy + 1; iiy++){
@@ -737,4 +489,27 @@ bool verlet_needs_update(){
 		}
 	}
 	return false;
+}
+void resize_cells(vector<disk> &particle, vector<vector<vector<int> > > &celllist){
+	cout << "Resizing the particle list and the cell lists...";
+	particle.resize(particles);
+	safe=particle;
+	Timesafe=0.0;
+	vnx = int(lx/verlet_grid);
+	vny = int(ly/verlet_grid);
+	if(vnx == 0){
+		vnx = 1;
+	}
+	if(vny == 0){
+		vny = 1;
+	}
+	dx = lx/vnx;
+	dy = ly/vny;
+
+	celllist.resize(vnx);
+	for(int i = 0; i < vnx; i++){
+		celllist[i].resize(vny);
+	}
+	//make_verlet();
+	cout << "OK\n";
 }
